App.py - initial file
App.py - initial data entry screen for the robo investing tool. The code contains some rudimentary methods for creating the portfolio. These were put in for testing purposes. For some reason some of the html code does not work properly on a smaller monitor. We will figure this out later.