Database changes
Create_table.py changed to add ID field as a primary key in some tables so that they can be linked when doing queries. Callbacks program comment query changed to be able to write record as the comment sqlite table changed. Reference program changed as we are no longer pulling data for S&P 500 spyders.