New files uploaded Sept 2, 2020
Changes have been made in the database. Auto-increment features have been installed and tables have been re-arranged so that the database is normalized.

Changes have been made to the login. The new code was taken from bootstrap.

Changes have been made to the employee transaction screen:
1) The date is now a single date datepicker
2) The customer name, transaction type and employee last name are now dropdowns

A rounding error was fixed in the sample portfolio.

The welcome screen now comes up when the system starts.